Fixed integer cast error

+++ diluvian/config.py
@@ -462,7 +462,7 @@ class Config(object):
         self.random_seed = int(settings.get("random_seed", 0))
         input_resolution = settings.get("input_resolution", None)
         if input_resolution is not None:
-            self.volume.resolution = [int(x) for x in input_resolution]
+            self.volume.resolution = [float(x) for x in input_resolution]
 
     def __str__(self):
         sanitized = {}
